% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrondCounts.R
\name{BrondCounts}
\alias{BrondCounts}
\title{BrondCounts}
\usage{
BrondCounts(data = c(), herz = c())
}
\arguments{
\item{data}{Matrix, with 3 columns with raw accelerometer data}

\item{herz}{numeric, sample rate of the data object}
}
\value{
no object is returned, only a new file is created in the output directory
}
\description{
BrondCounts
}
