% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrondCounts2csv.R
\name{BrondCounts2csv}
\alias{BrondCounts2csv}
\title{BrondCounts2csv}
\usage{
BrondCounts2csv(outputdir = c(), configfile = c())
}
\arguments{
\item{outputdir}{Path to output directory}

\item{configfile}{Path to configuration file}
}
\value{
no object is returned, only a new file is created in the output directory
}
\description{
BrondCounts2csv
}
