% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_params}
\alias{check_params}
\title{check_params}
\usage{
check_params(params = c())
}
\arguments{
\item{params}{Parameter object to be checked}
}
\value{
Data.frame with problematic parameter values and corresponding error
}
\description{
check_params
}
