% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConfigFile.R
\name{checkConfigFile}
\alias{checkConfigFile}
\title{Function to check that the config files have the expected format}
\usage{
checkConfigFile(file = c(), tool = c())
}
\arguments{
\item{file}{data path to config file}

\item{tool}{either PALMSpy, GGIR, or hbGIS for now.}
}
\value{
message with the result of the check (either ok or the description of a problem)
}
\description{
Function to check that the config files have the expected format
}
