% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFile.R
\name{checkFile}
\alias{checkFile}
\title{Function to check that the config files have the expected format}
\usage{
checkFile(file = c(), tool = c())
}
\arguments{
\item{file}{data path to config file}

\item{tool}{either PALMSpy, GGIR, or palmsplusr for now.}
}
\value{
message with the result of the check (either ok or the description of a problem)
}
\description{
Function to check that the config files have the expected format
}
