% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGGIRconfig.R
\name{checkGGIRconfig}
\alias{checkGGIRconfig}
\title{checkGGIRconfig}
\usage{
checkGGIRconfig(configfile = c())
}
\arguments{
\item{configfile}{Config object}
}
\value{
character with desiredtz value or error message
}
\description{
checkGGIRconfig
}
