% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPALMSpyconfig.R
\name{checkPALMSpyconfig}
\alias{checkPALMSpyconfig}
\title{checkPALMSpyconfig}
\usage{
checkPALMSpyconfig(configfile = c())
}
\arguments{
\item{configfile}{Config object}
}
\value{
character with desiredtz value or error message
}
\description{
checkPALMSpyconfig
}
