% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Counts2csv.R
\name{Counts2csv}
\alias{Counts2csv}
\title{Counts2csv}
\usage{
Counts2csv(outputdir = c(), configfile = c())
}
\arguments{
\item{outputdir}{Path to output directory}

\item{configfile}{Path to configuration file}
}
\value{
no object is returned, only a new file is created in the output directory
}
\description{
Counts2csv
}
