% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_files.R
\name{create_test_files}
\alias{create_test_files}
\title{create_test_files}
\usage{
create_test_files(dir = c(), Nfiles = 10, Nobs = 10)
}
\arguments{
\item{dir}{Directory where test files should be created}

\item{Nfiles}{Number of files to create}

\item{Nobs}{Number of observations}
}
\value{
no object is returned, function only writes file.
}
\description{
create_test_files
}
