% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_test_GGIRconfig.R
\name{create_test_GGIRconfig}
\alias{create_test_GGIRconfig}
\title{create_test_GGIRconfig}
\usage{
create_test_GGIRconfig(configfile = c(), desiredtz = "Europe/London")
}
\arguments{
\item{configfile}{Config object}

\item{desiredtz}{desiredtz to be used inside config file}
}
\value{
No object return only file created
}
\description{
create_test_GGIRconfig
}
