% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbGPS_shiny.R
\name{hbGPS_shiny}
\alias{hbGPS_shiny}
\title{hbGPS_shiny}
\usage{
hbGPS_shiny(
  ggiroutdir = NULL,
  gpsdir = NULL,
  outputdir = NULL,
  dataset_name = "",
  configfile = c()
)
}
\arguments{
\item{ggiroutdir}{GGIR time series output directory}

\item{gpsdir}{Path to GPS directory}

\item{outputdir}{Path to output directory}

\item{dataset_name}{Name of dataset}

\item{configfile}{Path to configuration file}
}
\value{
no object is returned, only a new file is created in the output directory
}
\description{
hbGPS_shiny
}
