% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbt_build_days.R
\name{hbt_build_days}
\alias{hbt_build_days}
\title{Calculate day-level summaries from the palmsplus dataset}
\usage{
hbt_build_days(
  data = NULL,
  verbose = TRUE,
  config_file = NULL,
  palmsplus_domains = NULL,
  domain_field = NULL,
  palmsplus_fields = NULL,
  home = NULL,
  school = NULL,
  home_nbh = NULL,
  school_nbh = NULL,
  participant_basis = NULL
)
}
\arguments{
\item{data}{The palmsplus data obtained from \code{\link{palms_build_palmsplus}}.}

\item{verbose}{Print progress to console. Default is \code{TRUE}.}

\item{config_file}{Path to the config file}

\item{palmsplus_domains}{...}

\item{domain_field}{...}

\item{palmsplus_fields}{...}

\item{home}{home}

\item{school}{school}

\item{home_nbh}{home_nbh}

\item{school_nbh}{school_nbh}

\item{participant_basis}{participant_basis}
}
\value{
A table summarised by day.
}
\description{
Build a days dataset by summarising \code{palmsplus}
by day and person (\code{identifier}). Not all variables in \code{palmsplus}
are summarised, only those specified using \code{\link{palms_add_field}} with
the argument \code{domain_field = TRUE}. By default, a \code{duration} field
is added (e.g., the total minutes per day).

All data are summarised by default. However, additional aggragation \emph{domains}
can be specified using \code{\link{palms_add_domain}} before building days.
Domains are a subset of data, such as during school time. All \code{domain_field}
variables will be summarised for each \emph{domain} seperatly.
}
