% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbt_build_multimodal.R
\name{hbt_build_multimodal}
\alias{hbt_build_multimodal}
\title{Build multimodal trips from trajectories}
\usage{
hbt_build_multimodal(
  data = NULL,
  spatial_threshold,
  temporal_threshold,
  palmsplus = NULL,
  verbose = TRUE,
  config_file = NULL,
  palmsplus_copy = NULL
)
}
\arguments{
\item{data}{The trajectories object built with \code{palms_calc_trajectories}.}

\item{spatial_threshold}{Spatial threshold in meters}

\item{temporal_threshold}{Temporal threshold in minutes}

\item{palmsplus}{...}

\item{verbose}{Print progress after each step. Default is \code{TRUE}.}

\item{config_file}{Path to the config file}

\item{palmsplus_copy}{When using config files, a copy of the palmsplus dataframe is passed to this function (future use)}
}
\value{
The input trajectories LINESTRING geometry, collapsed into multimodal trips
}
\description{
Build multimodal trips from trajectories.
}
\details{
Several columns are required in the \code{trajectories} dataset. These
need to be added as trajectory fields:
\itemize{
\item identifier
\item tripnumber
\item mot
\item start
\item end
\item geometry
}
}
