% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbt_build_palmsplus.R
\name{hbt_build_palmsplus}
\alias{hbt_build_palmsplus}
\title{Build the palmsplus dataset}
\usage{
hbt_build_palmsplus(
  data = NULL,
  verbose = TRUE,
  palmsplus_fields = NULL,
  loca = NULL,
  participant_basis = NULL
)
}
\arguments{
\item{data}{The PALMS data obtained using \code{\link{read_palms}}.}

\item{verbose}{Print progress to console after each iteration. Default is \code{TRUE}.}

\item{palmsplus_fields}{fields defined in PALMSplusRshiny}

\item{loca}{Nested list with location information}

\item{participant_basis}{participant_basis}
}
\description{
Build the \code{palmsplus} dataset by adding additional columns to the PALMS dataset.
The additional columns are specified using \code{\link{palms_add_field}}.
}
