% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbt_build_trajectories.R
\name{hbt_build_trajectories}
\alias{hbt_build_trajectories}
\title{Build trajectories from the palmsplus dataset}
\usage{
hbt_build_trajectories(
  data = NULL,
  trajectory_fields = NULL,
  trajectory_locations = NULL
)
}
\arguments{
\item{data}{The palmsplus data obtained from \code{\link{palms_build_palmsplus}}.}

\item{trajectory_fields}{trajectory_fields}

\item{trajectory_locations}{trajectory_locations}
}
\value{
A table of individual trips represented as \code{LINESTRING} geometry.
}
\description{
Build trajectories (trips) from the \code{palmsplus} dataset. This
returns a \code{sf data.frame} with \code{LINESTRING} geometry. Three columns
are returned by default (\code{identifier}, \code{tripnumber}, and \code{geometry}).
Additional columns can be specified with \code{\link{palms_add_trajectory_field}}
and \code{\link{palms_add_trajectory_location}}.
}
