% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbt_read_config.R
\name{hbt_read_config}
\alias{hbt_read_config}
\title{Read palmsplusr config file}
\usage{
hbt_read_config(path)
}
\arguments{
\item{path}{Path to config file}
}
\value{
dataframe
}
\description{
Read palmsplusr config file
}
