% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_tools.R
\name{identify_tools}
\alias{identify_tools}
\title{identify_tools}
\usage{
identify_tools(
  datatypes = c("AccRaw", "AccCount", "GPS", "GIS"),
  goals = c("PA", "Sleep", "QC", "Trips", "Environment"),
  available_tools = c("GGIR", "PALMS", "PALMSplus")
)
}
\arguments{
\item{datatypes}{Character vector with available data types}

\item{goals}{Character vector with research goals}

\item{available_tools}{Character vector with available tools/tools to consider}
}
\value{
List with proposed pipeline and tool input/ouptut properties
}
\description{
identify_tools
}
