% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_palmspy_params.R
\name{load_palmspy_params}
\alias{load_palmspy_params}
\title{load_palmspy_params}
\usage{
load_palmspy_params(file = c())
}
\arguments{
\item{file}{.json file}
}
\value{
list of parameters from the json object
}
\description{
load_palmspy_params
}
