% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modBuildPipelineUI.R
\name{modBuildPipelineUI}
\alias{modBuildPipelineUI}
\title{modBuildPipelineUI}
\usage{
modBuildPipelineUI(id)
}
\arguments{
\item{id}{...}
}
\value{
No object returned, this is a shiny module
}
\description{
modBuildPipelineUI
}
