% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modEditTable.R
\name{modEditTable}
\alias{modEditTable}
\title{smodEditTable}
\usage{
modEditTable(input, output, session, data, reset, save, configfile)
}
\arguments{
\item{input}{...}

\item{output}{...}

\item{session}{...}

\item{data}{...}

\item{reset}{...}

\item{save}{...}

\item{configfile}{...}
}
\value{
No object returned, this is a shiny module
}
\description{
smodEditTable
}
