% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modEditTableServer.R
\name{modEditTableServer}
\alias{modEditTableServer}
\title{modFunctionServer}
\usage{
modEditTableServer(id, reset, save, configfile)
}
\arguments{
\item{id}{...}

\item{reset}{...}

\item{save}{...}

\item{configfile}{...}
}
\value{
No object returned, this is a shiny module
}
\description{
modFunctionServer
}
