% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modEditTableUI.R
\name{modEditTableUI}
\alias{modEditTableUI}
\title{modEditTableUI}
\usage{
modEditTableUI(id)
}
\arguments{
\item{id}{...}
}
\value{
No object returned, this is a shiny module
}
\description{
modEditTableUI
}
