% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myApp.R
\name{myApp}
\alias{myApp}
\title{myApp}
\usage{
myApp(homedir = getwd(), envConda = "~/miniconda3/bin/conda", ...)
}
\arguments{
\item{homedir}{character to specify path to home directory}

\item{envConda}{character to specify path conda environment used for PALMSpy}

\item{...}{No input needed, function runs the app}
}
\value{
no object is returned, just an app
}
\description{
myApp
}
