% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRTool.R
\name{myRTool}
\alias{myRTool}
\title{myRTool}
\usage{
myRTool(inputdir, outputdir, config, desiredtz = c())
}
\arguments{
\item{inputdir}{Path to input directory}

\item{outputdir}{Path to output directory}

\item{config}{Config object}

\item{desiredtz}{Desired tz databasename}
}
\value{
no object is returned, only a new file is created in the output directory
}
\description{
myRTool
}
