% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palmsplus_build.R
\name{palmsplus_build}
\alias{palmsplus_build}
\title{Build the palmsplus objects}
\usage{
palmsplus_build(
  palms,
  config_file = NULL,
  verbose = TRUE,
  spatial_threshold,
  temporal_threshold,
  palmsplus_folder,
  dataset_name = "",
  palmsplus_fields,
  palmsplus_domains,
  trajectory_fields,
  multimodal_fields_def
)
}
\arguments{
\item{palms}{The PALMS data obtained using \code{\link{read_palms}}.}

\item{config_file}{Path to the config file}

\item{verbose}{Print progress to console after each iteration. Default is \code{TRUE}.}

\item{spatial_threshold}{Spatial threshold in meters}

\item{temporal_threshold}{Temporal threshold in minutes}

\item{palmsplus_folder}{path to direcotry where palmsplus results should be stored}

\item{dataset_name}{...}

\item{palmsplus_fields}{...}

\item{palmsplus_domains}{...}

\item{trajectory_fields}{...}

\item{multimodal_fields_def}{...}
}
\description{
Build the \code{palmsplus} objects by adding additional columns..
}
