% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palmsplusr_shiny.R
\name{palmsplusr_shiny}
\alias{palmsplusr_shiny}
\title{palmsplusr_shiny}
\usage{
palmsplusr_shiny(
  gisdir = "",
  palmsdir = "",
  gislinkfile = "",
  outputdir = "",
  dataset_name = "",
  configfile = ""
)
}
\arguments{
\item{gisdir}{Path to directory with GIS files}

\item{palmsdir}{Path to PALMSpy or PALMS output directory}

\item{gislinkfile}{Path to participant basis file, which is the file that links all participant identifies with the GIS data}

\item{outputdir}{Path to outputdir location}

\item{dataset_name}{Name of dataset}

\item{configfile}{Configuration file}
}
\value{
palms_to_clean_lower object
}
\description{
palmsplusr_shiny
}
