% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PALMSplusRshiny.R
\name{PALMSplusRshiny}
\alias{PALMSplusRshiny}
\title{PALMSplusRshiny}
\usage{
PALMSplusRshiny(gisdir = "", palmsdir = "", gislinkfile = "", outputdir = "")
}
\arguments{
\item{gisdir}{Path to directory with GIS files}

\item{palmsdir}{Path to PALMSpy or PALMS output directory}

\item{gislinkfile}{Path to participant basis file, which is the file that links all participant identifies with the GIS data}

\item{outputdir}{Path to outputdir location}
}
\value{
palms_to_clean_lower object
}
\description{
PALMSplusRshiny
}
