% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PALMSpy_R.R
\name{PALMSpy_R}
\alias{PALMSpy_R}
\title{PALMSpy_R}
\usage{
PALMSpy_R(gps_path, acc_path, config_file = c())
}
\arguments{
\item{gps_path}{Path to gps data}

\item{acc_path}{Path to acc data}

\item{config_file}{Path to configuration file}
}
\value{
No object is returned, function only writes message to console and writes file.
}
\description{
PALMSpy_R
}
