% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PALMSpyshiny.R
\name{PALMSpyshiny}
\alias{PALMSpyshiny}
\title{PALMSpyshiny}
\usage{
PALMSpyshiny(outputdir, gpsdir, count_file_location)
}
\arguments{
\item{outputdir}{Path to output directory}

\item{gpsdir}{Path to GPS files}

\item{count_file_location}{Path to count files}
}
\value{
no object is returned, only a new file is created in the output directory
}
\description{
PALMSpyshiny
}
