% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runpalmsplus.R
\name{runpalmsplus}
\alias{runpalmsplus}
\title{runpalmsplus}
\usage{
runpalmsplus(gisdir = "", palmsdir = "", gislinkfile = "")
}
\arguments{
\item{gisdir}{Path to directory with GIS files}

\item{palmsdir}{Path to PALMSpy or PALMS output directory}

\item{gislinkfile}{Path to participant basis file, which is the file that links all participant identifies with the GIS data}
}
\value{
palms_to_clean_lower object
}
\description{
runpalmsplus
}
