% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_palmspy_params.R
\name{save_palmspy_params}
\alias{save_palmspy_params}
\title{save_palmspy_params}
\usage{
save_palmspy_params(new_params = c(), file = c())
}
\arguments{
\item{new_params}{new parameters}

\item{file}{.json file}
}
\value{
No object returned, function only wrtite data
}
\description{
save_palmspy_params
}
