% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_tools.R
\docType{class}
\name{toolio-class}
\alias{toolio-class}
\title{identify_tools}
\arguments{
\item{datatypes}{Character vector with available data types}

\item{goals}{Character vector with research goals}

\item{available_tools}{Character vector with available tools/tools to consider}
}
\value{
List with proposed pipeline and tool input/ouptut properties
}
\description{
identify_tools
}
