% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGGIRconfig.R
\name{updateGGIRconfig}
\alias{updateGGIRconfig}
\title{updateGGIRconfig}
\usage{
updateGGIRconfig(configfile = c(), new_desiredtz)
}
\arguments{
\item{configfile}{Config object}

\item{new_desiredtz}{New desiredtz value}
}
\value{
No object return only file updated
}
\description{
updateGGIRconfig
}
