% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePALMSpyconfig.R
\name{updatePALMSpyconfig}
\alias{updatePALMSpyconfig}
\title{updatePALMSpyconfig}
\usage{
updatePALMSpyconfig(configfile = c(), new_desiredtz)
}
\arguments{
\item{configfile}{Config object}

\item{new_desiredtz}{New desiredtz value}
}
\value{
No object return only file updated
}
\description{
updatePALMSpyconfig
}
