% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGIRshiny.R
\name{GGIRshiny}
\alias{GGIRshiny}
\title{GGIR}
\usage{
GGIRshiny(inputdir, outputdir, config, sleepdiary)
}
\arguments{
\item{inputdir}{Path to input directory}

\item{outputdir}{Path to output directory}

\item{config}{Config object}

\item{sleepdiary}{Path to sleep diary}
}
\value{
no object is returned, only a new file is created in the output directory
}
\description{
GGIR
}
