% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGIRshiny.R
\name{GGIRshiny}
\alias{GGIRshiny}
\title{GGIRshiny}
\usage{
GGIRshiny(rawaccdir, outputdir, configfile = c(), sleepdiary = c())
}
\arguments{
\item{rawaccdir}{Path to input directory}

\item{outputdir}{Path to output directory}

\item{configfile}{Configfile path}

\item{sleepdiary}{Path to sleep diary}
}
\value{
no object is returned, only a new file is created in the output directory
}
\description{
GGIRshiny
}
