% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbt_check_missing_id.R
\name{hbt_check_missing_id}
\alias{hbt_check_missing_id}
\title{hbt_check_missing_id}
\usage{
hbt_check_missing_id(
  participant_basis,
  palmsplus_folder,
  dataset_name,
  palms,
  loca,
  groupinglocation = "school",
  verbose = TRUE
)
}
\arguments{
\item{participant_basis}{object as loaded inside PALMSplusRshiny}

\item{palmsplus_folder}{output folder path for storing log of excluded IDs}

\item{dataset_name}{Name of dataset}

\item{palms}{palms object as loaded inside PALMSplusRshiny}

\item{loca}{Nested list with location information}

\item{groupinglocation}{groupinglocation}

\item{verbose}{verbose}
}
\value{
List with participant_basis and palms object without non-matching IDs
}
\description{
hbt_check_missing_id
}
