% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_params.R
\name{load_params}
\alias{load_params}
\title{load_params}
\usage{
load_params(file = c(), format = "json_palmspy")
}
\arguments{
\item{file}{Character to specify location of configuration file}

\item{format}{Character to specify format of configuration file: json_palsmpy or csv_GGIR}
}
\value{
list of parameters extract from the configuration file
}
\description{
load_params
}
