% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modBuildPipelineServer.R
\name{modBuildPipelineServer}
\alias{modBuildPipelineServer}
\title{modBuildPipelineServer}
\usage{
modBuildPipelineServer(id)
}
\arguments{
\item{id}{...}
}
\value{
List of pipeline, availabledata and research goals
}
\description{
modBuildPipelineServer
}
