% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modConfigServer.R
\name{modConfigServer}
\alias{modConfigServer}
\title{modConfigServer}
\usage{
modConfigServer(id, reset, save, configfile, tool)
}
\arguments{
\item{id}{...}

\item{reset}{...}

\item{save}{...}

\item{configfile}{...}

\item{tool}{...}
}
\value{
No object returned, this is a shiny module
}
\description{
modConfigServer
}
