% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modConfigServer.R
\name{modConfigServer}
\alias{modConfigServer}
\title{modConfigServer}
\usage{
modConfigServer(id, tool, homedir = getwd(), prevConfig = c())
}
\arguments{
\item{id}{...}

\item{tool}{...}

\item{homedir}{character to specify home directory}

\item{prevConfig}{character to specify path to config file selected in previous run}
}
\value{
No object returned, this is a shiny module
}
\description{
modConfigServer
}
