% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myApp.R
\name{myApp}
\alias{myApp}
\title{myApp}
\usage{
myApp(homedir = getwd(), ...)
}
\arguments{
\item{homedir}{character to specify home directory}

\item{...}{No input needed, function runs the app}
}
\value{
no object is returned, just an app
}
\description{
myApp
}
