% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPyTool.R
\name{myPyTool}
\alias{myPyTool}
\title{myPyTool}
\usage{
myPyTool(inputdir, outputdir, sleepdiary)
}
\arguments{
\item{inputdir}{Path to input directory}

\item{outputdir}{Path to output directory}
}
\value{
no object is returned, function only writes message to console and writes file.
}
\description{
myPyTool
}
