% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRTool.R
\name{myRTool}
\alias{myRTool}
\title{myRTool}
\usage{
myRTool(accdir, outputdir, config = c())
}
\arguments{
\item{accdir}{Path to input directory}

\item{outputdir}{Path to output directory}

\item{config}{Config object}
}
\value{
no object is returned, only a new file is created in the output directory
}
\description{
myRTool
}
